Feat/asl citizen videomaev2 demo baseline

+++ src/silent_signal/cli/prepare_graph.py
@@ -295,7 +295,7 @@ def _mean(values: Sequence[float]) -> float | None:
 
 
 def _duration(seconds: float) -> str:
-    total = max(0, int(round(seconds)))
+    total = max(0, round(seconds))
     hours, remainder = divmod(total, 3600)
     minutes, secs = divmod(remainder, 60)
     return f"{hours:02d}:{minutes:02d}:{secs:02d}"
